+++ docs/HISTORY.rst
@@ -4,8 +4,8 @@ Changelog
 0.6.0 (unreleased)
 ------------------
 
-- Plone 5 compatibility
-  [cekk]
+- Plone 5 compatibility.
+  [cekk, mamico]
 
 
 0.5.0 (2017-01-20)